cleanup: rename string utils file
