+++ src/components/HeaderSearch.vue
@@ -3,7 +3,7 @@ import { defineComponent, ref, watch } from 'vue';
 import { useRouter } from 'vue-router';
 import { OptionsObj } from 'src/types';
 import { api } from 'src/api';
-import { isValidTransactionHex } from 'src/utils/stringValidator';
+import { isValidTransactionHex } from 'src/utils/string-utils';
 import { useQuasar } from 'quasar';
 
 export default defineComponent({
